+++ src/test/java/gift/controller/ProductControllerTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.doNothing;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import gift.dto.MemberDTO;
+import gift.dto.betweenClient.member.MemberDTO;
 import gift.service.MemberService;
 import gift.service.ProductService;
 import gift.util.JwtUtil;
